refactor: added PostCardFooter component
